+++ db.h
@@ -100,6 +100,8 @@ namespace DB {
         std::wstring name;
         std::wstring exe_path;
         std::wstring working_dir;
+        std::wstring arguments;
+        std::wstring comment;
         std::wstring icon_path;
         int icon_index = 0;
         int run_as_admin = 0;